requirements.txt
py-cord[voice]==1.7.3
discord==1.7.3
discord-py-slash-command==3.0.3
mysql-connector-python==8.0.27
python-dotenv==0.19.2
python-version==0.0.2
markdownify==0.10.1
youtube_dl==2021.12.17